+++ src/chart/rendering/drawinterlacing.cpp
@@ -133,32 +133,35 @@ void DrawInterlacing::draw(
 		auto textAlpha =
 		    Math::FuzzyBool::And<double>(weight, enabled.labels);
 
-		if (std::signbit(rangeSize) != std::signbit(stepSize)
-		    || Math::Floating::is_zero(rangeSize))
-			return;
+		auto singleLabelRange = Math::Floating::is_zero(rangeSize);
 
-		auto stripWidth = stepSize / rangeSize;
+		double stripWidth{};
+		if (singleLabelRange)
+			stepSize = 1.0;
+		else {
+			stripWidth = stepSize / rangeSize;
+			if (stripWidth <= 0) return;
+		}
 
 		auto axisBottom = axis.origo() + stripWidth;
 
 		auto iMin =
-		    axisBottom > 0 ? static_cast<int>(
-		        std::floor(-axis.origo() / (2 * stripWidth)))
-		                   : static_cast<int>(
-		                       (axis.range.getMin() - stepSize) / 2);
+		    axisBottom > 0
+		        ? std::floor(-axis.origo() / (2 * stripWidth)) * 2
+		        : std::round(axis.range.getMin() - stepSize);
 
-		if (stripWidth <= 0) return;
 		auto interlaceCount = 0U;
 		const auto maxInterlaceCount = 1000U;
-		for (int i = iMin; ++interlaceCount <= maxInterlaceCount;
-		     ++i) {
-			auto bottom = axisBottom + i * 2 * stripWidth;
-			if (bottom >= 1.0) break;
+		for (auto i = static_cast<int>(iMin);
+		     ++interlaceCount <= maxInterlaceCount;
+		     i += 2) {
+			auto bottom = axisBottom + i * stripWidth;
+			if (bottom > 1.0) break;
 			auto clippedBottom = bottom;
 			auto top = bottom + stripWidth;
 			auto clipTop = top > 1.0;
 			auto clipBottom = bottom < 0.0;
-			auto topUnderflow = top <= 0.0;
+			auto topUnderflow = top < 0.0;
 			if (clipTop) top = 1.0;
 			if (clipBottom) clippedBottom = 0.0;
 
@@ -174,7 +177,7 @@ void DrawInterlacing::draw(
 					canvas.setFont(Gfx::Font{axisStyle.label});
 
 					if (!clipBottom) {
-						auto value = (i * 2 + 1) * stepSize;
+						auto value = (i + 1) * stepSize;
 						auto tickPos =
 						    rect.bottomLeft().comp(!horizontal)
 						    + origo.comp(horizontal);
@@ -192,8 +195,9 @@ void DrawInterlacing::draw(
 							    horizontal,
 							    tickPos);
 					}
+					if (singleLabelRange) break;
 					if (!clipTop) {
-						auto value = (i * 2 + 2) * stepSize;
+						auto value = (i + 2) * stepSize;
 						auto tickPos =
 						    rect.topRight().comp(!horizontal)
 						    + origo.comp(horizontal);
@@ -212,7 +216,7 @@ void DrawInterlacing::draw(
 							    tickPos);
 					}
 				}
-				else {
+				else if (!singleLabelRange) {
 					canvas.save();
 
 					canvas.setLineColor(Gfx::Color::Transparent());