+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix axis interpolation. From now the axis and axis labels are following the markers.
 - Fix measure axis labels when the range started after the 2000th step value from origo. 
 - Remove marker labels at intermediate steps.
+- Fix single value range interpolation and show.
 - Markers are the same even if new record added.
 
 ### Added